refactor: 계산에만 집중할 수 있도록 역할 분리
